feat: remake field parser, touch httpStatus.hpp ~
