+++ .gitignore
@@ -2,7 +2,6 @@
 
 # executive files
 webserv
-test.out
 
 # Prerequisites
 *.d
@@ -36,3 +35,6 @@ test.out
 *.exe
 *.out
 *.app
+
+# log files
+*.log
