+++ test/http/request/Makefile
@@ -24,6 +24,7 @@ clean:
 
 fclean: clean
 	$(RM) $(NAME)
+	$(RM) field_test.log
 
 re: fclean all
