+++ test/http/request/test_main.cpp
@@ -2,8 +2,14 @@
 
 #include "../../../toolbox/stepmark.hpp"
 
+namespace http {
 void fieldTest();
+}
 
+// グローバルスコープのmain関数
 int main(void) {
-    fieldTest();
+    toolbox::logger::StepMark::setLogFile("field_test.log");
+    toolbox::logger::StepMark::setLevel(toolbox::logger::DEBUG);
+    http::fieldTest();
+    return 0;
 }
